+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,31 +1,39 @@
-import NextAuth from "next-auth"
-import CredentialsProvider from "next-auth/providers/credentials"
+import bcrypt from 'bcrypt';
+import prisma from '@/prisma';
+import NextAuth from 'next-auth';
+import CredentialsProvider from 'next-auth/providers/credentials';
 
-export const authOptions = {
+const handler = NextAuth({
+    secret: process.env.AUTH_SECRET,
     providers: [
         CredentialsProvider({
             name: 'Credentials',
             credentials: {
-                username: { label: "Username", type: "text", placeholder: "jsmith" },
-                password: { label: "Password", type: "password" }
+                email: { label: 'Email', type: 'text' },
+                password: { label: 'Password', type: 'password' },
             },
-            async authorize(credentials, req) {
-                const res = await fetch("/your/endpoint", {
-                    method: 'POST',
-                    body: JSON.stringify(credentials),
-                    headers: { "Content-Type": "application/json" }
-                })
-                const user = await res.json()
+            async authorize(credentials, _req) {
+                if (!credentials?.email || !credentials?.password) return null;
 
-                if (res.ok && user) {
-                    return user
-                }
-                return null
-            }
-        })
+                const { email, password } = credentials;
+
+                const user = await prisma.user.findUnique({ where: { email } });
+                if (!user) return null;
+
+                const passwordMatch = await bcrypt.compare(password, user.password);
+                if (!passwordMatch) return null;
+
+                return {
+                    ...user,
+                    id: user.id.toString(),
+                    name: `${user.firstName} ${user.lastName}`,
+                };
+            },
+        }),
     ],
     pages: {
-        signIn: '/login'
-    }
-}
-export default NextAuth(authOptions)
+        signIn: '/login',
+    },
+});
+
+export { handler as GET, handler as POST }