+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Handle messages in signup form using 'react-tostify'
 - Add src folder
 - Add ui folder into components folder
+- Implement Credentials login using Next-auth
 
 ## Sprint 1 (November) [01 Nov, 2023 - 10 Nov, 2023]
 ### [released]